+++ src/fundamentals/wire.rs
@@ -129,7 +129,7 @@ pub struct TimeseriesData {
     pub(crate) timestamp: Option<Vec<i64>>,
     #[serde(default)]
     #[allow(dead_code)]
-    meta: serde_json::Value,
+    pub(crate) meta: serde_json::Value,
     #[serde(flatten)]
     pub(crate) values: std::collections::HashMap<String, serde_json::Value>,
 }